feat: move hashtag, geohash, chess filter assemblers to commons
Move 3 complete filter assembler chains to commons KMP module:
- Hashtag: HashtagQueryState, HashtagFilterAssembler, HashtagFeedFilterSubAssembler, FilterPostsByHashtags
- Geohash: GeohashQueryState, GeoHashFilterAssembler, GeoHashFeedFilterSubAssembler, FilterPostsByGeohash
- Chess: ChessQueryState, ChessFilterAssembler, ChessFeedFilterSubAssembler, FilterChessEvent

Also moved CommentKinds to commons as shared constant.

SubAssemblers migrated from PerUniqueIdEoseManager to PerKeyEoseManager
(the KMP-compatible equivalent already in commons).

Amethyst files replaced with typealias re-exports for backwards compatibility.
Build verified: commons:compileKotlinJvm + amethyst:compilePlayDebugKotlin pass.
