+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/chess/datasource/ChessFeedFilterSubAssembler.kt
@@ -20,67 +20,6 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.chess.datasource
 
-import com.vitorpamplona.amethyst.service.relayClient.eoseManagers.PerUniqueIdEoseManager
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.nip01Core.core.Event
-import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.client.reqs.SubscriptionListener
-import com.vitorpamplona.quartz.nip01Core.relay.client.subscriptions.Subscription
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.utils.TimeUtils
-
-/**
- * Query state for chess subscription
- */
-data class ChessQueryState(
-    val userPubkey: String,
-    val inboxRelays: Set<NormalizedRelayUrl>,
-    val globalRelays: Set<NormalizedRelayUrl>,
-    val isGlobal: Boolean,
-    val activeGameIds: Set<String> = emptySet(),
-    val opponentPubkeys: Set<String> = emptySet(),
-)
-
-/**
- * Sub-assembler that creates the actual relay filters.
- * Intercepts incoming chess events from relays and forwards them to the ViewModel.
- */
-class ChessFeedFilterSubAssembler(
-    client: INostrClient,
-    allKeys: () -> Set<ChessQueryState>,
-) : PerUniqueIdEoseManager<ChessQueryState, String>(client, allKeys) {
-    var onChessEvent: ((Event) -> Unit)? = null
-
-    override fun updateFilter(
-        key: ChessQueryState,
-        since: SincePerRelayMap?,
-    ): List<RelayBasedFilter> = filterChessEvents(key, since)
-
-    override fun id(key: ChessQueryState): String = key.userPubkey
-
-    override fun newSub(key: ChessQueryState): Subscription =
-        requestNewSubscription(
-            object : SubscriptionListener {
-                override fun onEose(
-                    relay: NormalizedRelayUrl,
-                    forFilters: List<Filter>?,
-                ) {
-                    newEose(key, relay, TimeUtils.now(), forFilters)
-                }
-
-                override fun onEvent(
-                    event: Event,
-                    isLive: Boolean,
-                    relay: NormalizedRelayUrl,
-                    forFilters: List<Filter>?,
-                ) {
-                    if (isLive) {
-                        newEose(key, relay, TimeUtils.now(), forFilters)
-                    }
-                    onChessEvent?.invoke(event)
-                }
-            },
-        )
-}
+// Re-export from commons for backwards compatibility
+typealias ChessQueryState = com.vitorpamplona.amethyst.commons.relayClient.assemblers.chess.ChessQueryState
+typealias ChessFeedFilterSubAssembler = com.vitorpamplona.amethyst.commons.relayClient.assemblers.chess.ChessFeedFilterSubAssembler
