+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/chess/datasource/FilterChessEvent.kt
@@ -18,41 +18,8 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
+@file:Suppress("unused")
+
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.chess.datasource
 
-import com.vitorpamplona.amethyst.commons.chess.subscription.ChessFilterBuilder
-import com.vitorpamplona.amethyst.commons.chess.subscription.ChessSubscriptionState
-import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-
-/**
- * Create relay filters for chess events using the shared [ChessFilterBuilder].
- *
- * Following jesterui's approach:
- * 1. Fetch ALL challenges from ALL connected relays (no author/tag restriction)
- * 2. Filter client-side based on user's preferences (Global/Follows)
- * 3. Game-specific subscriptions for active games
- *
- * This ensures we see:
- * - Open challenges from anyone
- * - Challenges directed at us
- * - Public games to spectate
- */
-fun filterChessEvents(
-    key: ChessQueryState,
-    since: SincePerRelayMap?,
-): List<RelayBasedFilter> {
-    // Convert Android ChessQueryState to shared ChessSubscriptionState
-    val state =
-        ChessSubscriptionState(
-            userPubkey = key.userPubkey,
-            relays = key.inboxRelays + key.globalRelays,
-            activeGameIds = key.activeGameIds,
-            opponentPubkeys = key.opponentPubkeys,
-        )
-
-    // Use shared filter builder for consistent behavior with Desktop
-    return ChessFilterBuilder.buildAllFilters(state) { relay ->
-        since?.get(relay)?.time
-    }
-}
+// Re-export from commons for backwards compatibility
