+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/geohash/datasource/GeoHashFeedFilterSubAssembler.kt
@@ -20,22 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.geohash.datasource
 
-import com.vitorpamplona.amethyst.service.relayClient.eoseManagers.PerUniqueIdEoseManager
-import com.vitorpamplona.amethyst.service.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-
-class GeoHashFeedFilterSubAssembler(
-    client: INostrClient,
-    allKeys: () -> Set<GeohashQueryState>,
-) : PerUniqueIdEoseManager<GeohashQueryState, String>(client, allKeys) {
-    override fun updateFilter(
-        key: GeohashQueryState,
-        since: SincePerRelayMap?,
-    ): List<RelayBasedFilter> = filterPostsByGeohash(key.geohash, key.relays, since)
-
-    /**
-     * Only one key per hashtag.
-     */
-    override fun id(key: GeohashQueryState) = key.lowercaseGeohash
-}
+// Re-export from commons for backwards compatibility
+typealias GeoHashFeedFilterSubAssembler = com.vitorpamplona.amethyst.commons.relayClient.assemblers.geohash.GeoHashFeedFilterSubAssembler
