+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/geohash/datasource/GeoHashFilterAssembler.kt
@@ -20,34 +20,6 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.geohash.datasource
 
-import androidx.compose.runtime.Stable
-import com.vitorpamplona.amethyst.commons.relayClient.composeSubscriptionManagers.ComposeSubscriptionManager
-import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-
-// This allows multiple screen to be listening to tags, even the same tag
-class GeohashQueryState(
-    val geohash: String,
-    val relays: Set<NormalizedRelayUrl>,
-) {
-    val lowercaseGeohash = geohash.lowercase()
-}
-
-/**
- * Creates a filter for multiple geohashes at the same time.
- */
-@Stable
-class GeoHashFilterAssembler(
-    client: INostrClient,
-) : ComposeSubscriptionManager<GeohashQueryState>() {
-    val group =
-        listOf(
-            GeoHashFeedFilterSubAssembler(client, ::allKeys),
-        )
-
-    override fun invalidateKeys() = invalidateFilters()
-
-    override fun invalidateFilters() = group.forEach { it.invalidateFilters() }
-
-    override fun destroy() = group.forEach { it.destroy() }
-}
+// Re-export from commons for backwards compatibility
+typealias GeohashQueryState = com.vitorpamplona.amethyst.commons.relayClient.assemblers.geohash.GeohashQueryState
+typealias GeoHashFilterAssembler = com.vitorpamplona.amethyst.commons.relayClient.assemblers.geohash.GeoHashFilterAssembler
