+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/relayClient/assemblers/chess/ChessFilterAssembler.kt
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.relayClient.assemblers.chess
+
+import com.vitorpamplona.amethyst.commons.relayClient.composeSubscriptionManagers.ComposeSubscriptionManager
+import com.vitorpamplona.quartz.nip01Core.core.Event
+import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
+import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
+
+/**
+ * Query state for chess subscription
+ */
+data class ChessQueryState(
+    val userPubkey: String,
+    val inboxRelays: Set<NormalizedRelayUrl>,
+    val globalRelays: Set<NormalizedRelayUrl>,
+    val isGlobal: Boolean,
+    val activeGameIds: Set<String> = emptySet(),
+    val opponentPubkeys: Set<String> = emptySet(),
+)
+
+/**
+ * Filter assembler for chess events
+ */
+class ChessFilterAssembler(
+    client: INostrClient,
+) : ComposeSubscriptionManager<ChessQueryState>() {
+    private val subAssembler = ChessFeedFilterSubAssembler(client, ::allKeys)
+
+    val group = listOf(subAssembler)
+
+    fun setOnChessEvent(callback: ((Event) -> Unit)?) {
+        subAssembler.onChessEvent = callback
+    }
+
+    override fun invalidateKeys() = invalidateFilters()
+
+    override fun invalidateFilters() = group.forEach { it.invalidateFilters() }
+
+    override fun destroy() = group.forEach { it.destroy() }
+}
