+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/relayClient/assemblers/geohash/GeoHashFeedFilterSubAssembler.kt
@@ -18,8 +18,24 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.hashtag.datasource
+package com.vitorpamplona.amethyst.commons.relayClient.assemblers.geohash
 
-// Re-export from commons for backwards compatibility
-typealias HashtagQueryState = com.vitorpamplona.amethyst.commons.relayClient.assemblers.hashtag.HashtagQueryState
-typealias HashtagFilterAssembler = com.vitorpamplona.amethyst.commons.relayClient.assemblers.hashtag.HashtagFilterAssembler
+import com.vitorpamplona.amethyst.commons.relayClient.eoseManagers.PerKeyEoseManager
+import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
+import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
+import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
+
+class GeoHashFeedFilterSubAssembler(
+    client: INostrClient,
+    allKeys: () -> Set<GeohashQueryState>,
+) : PerKeyEoseManager<GeohashQueryState, String>(client, allKeys) {
+    override fun updateFilter(
+        queryState: GeohashQueryState,
+        since: SincePerRelayMap?,
+    ): List<RelayBasedFilter> = filterPostsByGeohash(queryState.geohash, queryState.relays, since)
+
+    /**
+     * Only one key per geohash.
+     */
+    override fun extractKey(queryState: GeohashQueryState) = queryState.lowercaseGeohash
+}
